Move inner ADLImporter classes up one level
