+++ org.hl7.fhir.convertors/src/main/java/org/hl7/fhir/convertors/misc/adl/Cardinality.java
@@ -1,0 +1,9 @@
+package org.hl7.fhir.convertors.misc.adl;
+
+import lombok.Data;
+
+@Data
+public class Cardinality {
+  private String min;
+  private String max;
+}
