+++ org.hl7.fhir.convertors/src/main/java/org/hl7/fhir/convertors/misc/adl/TextSet.java
@@ -1,0 +1,11 @@
+package org.hl7.fhir.convertors.misc.adl;
+
+import lombok.Data;
+
+@Data
+public class TextSet {
+
+  private String text;
+  private String description;
+  private String comment;
+}
